+++ src/components/reusedComponents/Form.js
@@ -1,21 +1,41 @@
 import React from "react";
-import { handleFormInput } from "../../api/fetch";
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
+import { handleFormInput, createMedia, newFormSubmitHandle } from "../../api/fetch";
+import "../shows/ShowsForm.css"
 
 function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoint, navigateVar, paramVar}) {
+  // Declare state for obj shapefor input
+  const [input, setInput] = useState({
+    type: "",
+    title: "",
+    country: "",
+    dateAdded: "",
+    description: "",
+    duration: "",
+    listedIn: "",
+    rating: "",
+    releaseYear: ""
+  })
+
+  // Declare variable for navigate -> send back to index page
+  const navigate = useNavigate()
+
+
   return (
     <form
       onSubmit={(event) => {!paramVar ?
-        submitFunction(event, stateVar, endpoint, navigateVar, fetchFunction) :
-        submitFunction(event, stateVar, endpoint, navigateVar, fetchFunction, paramVar)
+        newFormSubmitHandle(event, input, endpoint, navigate, createMedia) :
+        newFormSubmitHandle(event, stateVar, endpoint, navigateVar, fetchFunction, paramVar)
       }}
     >
       <label htmlFor="title">Title:</label>
       <input
         type="text"
         id="title"
-        value={stateVar.title}
+        value={input.title}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -23,9 +43,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="description"
-        value={stateVar.description}
+        value={input.description}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -33,9 +53,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="type"
-        value={stateVar.type}
+        value={input.type}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -43,9 +63,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="rating"
-        value={stateVar.rating}
+        value={input.rating}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -53,9 +73,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="listedIn"
-        value={stateVar.listedIn}
+        value={input.listedIn}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -63,9 +83,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="duration"
-        value={stateVar.duration}
+        value={input.duration}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -73,9 +93,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="releaseYear"
-        value={stateVar.releaseYear}
+        value={input.releaseYear}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -83,9 +103,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="country"
-        value={stateVar.country}
+        value={input.country}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
 
@@ -93,9 +113,9 @@ function Form({submitFunction, fetchFunction, stateVar, setStateFunction, endpoi
       <input
         type="text"
         id="dateAdded"
-        value={stateVar.dateAdded}
+        value={input.dateAdded}
         onChange={(event) => {
-          handleFormInput(event, stateVar, setStateFunction);
+          handleFormInput(event, input, setInput);
         }}
       />
